Improve GUI and code quality
The GUI interface cannot resize and the text does not wrap around
properly.

The parse method in Parser is long and could become unwieldy for future
iterations and updates.

Let's,
* Update the properties of dialogbox.fxml in scenebuilder
* Rename the application to ManMeowth for personality
* Extract methods from within the parse method
