--- data/duke.txt
@@ -1,3 +1,0 @@
-T | 1 | 1
-T | 0 | hello
-T | 0 | help
