+++ data/man_meowth.txt
@@ -1,0 +1,8 @@
+T | 1 | 5
+T | 0 | 6
+T | 0 | 7
+T | 1 | 8
+T | 0 | 9
+T | 0 | hello
+D | 1 | hi | tonight
+E | 1 | abc | 2022-12-121982
