--- src/main/java/duke/Launcher.java
@@ -1,17 +1,0 @@
-package duke;
-
-import javafx.application.Application;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-public class Launcher {
-    /**
-     * Provides entry point into Duke program using GUI.
-     *
-     * @param args The command line arguments.
-     **/
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
