--- src/main/java/duke/command/MarkCommand.java
@@ -1,41 +1,0 @@
-package duke.command;
-
-import duke.exception.DukeException;
-import duke.list.TaskList;
-import duke.storage.ListLoader;
-import duke.task.Task;
-import duke.ui.Ui;
-
-/**
- * Represents command to Duke to mark a task as complete.
- *
- * @author WR3nd3
- */
-public class MarkCommand extends Command {
-    /** Position of task to be marked as complete in the list */
-    private final int position;
-
-    /**
-     * Constructs MarkCommand for a given task.
-     *
-     * @param position Integer representing position of task to be marked as complete in the list.
-     */
-    public MarkCommand(int position) {
-        this.position = position;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String execute(TaskList tasks, Ui ui, ListLoader storage) throws DukeException {
-        if (tasks.isValidPosition(position)) {
-            Task t = tasks.retrieveRank(position);
-            storage.markTask(t.summary());
-            tasks.mark(position);
-            return ui.showMark(t);
-        } else {
-            return ui.showEmpty();
-        }
-    }
-}
