--- src/main/java/duke/task/TaskId.java
@@ -1,10 +1,0 @@
-package duke.task;
-
-/**
- * Identifiers to be used for tasks.
- *
- * @author WR3nd3
- */
-public enum TaskId {
-    T, E, D
-}
