--- src/main/java/duke/ui/DialogBox.java
@@ -1,81 +1,0 @@
-package duke.ui;
-
-import java.io.IOException;
-import java.util.Collections;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.HBox;
-
-/**
- * An example of a custom control using FXML.
- * This control represents a dialog box consisting of an ImageView to represent the speaker's face and a label
- * containing text from the speaker.
- */
-public class DialogBox extends HBox {
-    @FXML
-    private Label dialog;
-    @FXML
-    private ImageView displayPicture;
-
-    /**
-     * Constructor for the DialogBox object.
-     *
-     * @param text The text to be displayed in the DialogBox.
-     * @param img The image to be displayed in the DialogBox.
-     */
-    private DialogBox(String text, Image img) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
-            fxmlLoader.setController(this);
-            fxmlLoader.setRoot(this);
-            fxmlLoader.load();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        dialog.setText(text);
-        displayPicture.setImage(img);
-    }
-
-    /**
-     * Flips the dialog box such that the ImageView is on the left and text on the right.
-     */
-    private void flip() {
-        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-        Collections.reverse(tmp);
-        getChildren().setAll(tmp);
-        setAlignment(Pos.TOP_LEFT);
-    }
-
-    /**
-     * Returns DialogBox for the Duke program's response.
-     *
-     * @param text Duke's text response to the user.
-     * @param img Duke's avatar image.
-     * @return DialogBox for the Duke program's response.
-     */
-    public static DialogBox getDukeDialog(String text, Image img) {
-        return new DialogBox(text, img);
-    }
-
-    /**
-     * Returns DialogBox for the user's input.
-     *
-     * @param text The user's input text.
-     * @param img User's avatar image.
-     * @return Returns DialogBox for the user's input.
-     */
-    public static DialogBox getUserDialog(String text, Image img) {
-        var db = new DialogBox(text, img);
-        db.flip();
-        return db;
-    }
-}
