--- src/main/java/duke/ui/MainWindow.java
@@ -1,68 +1,0 @@
-package duke.ui;
-
-import duke.Duke;
-import javafx.application.Platform;
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.VBox;
-/**
- * Controller for MainWindow. Provides the layout for the other controls.
- */
-public class MainWindow extends AnchorPane {
-    @FXML
-    private ScrollPane scrollPane;
-    @FXML
-    private VBox dialogContainer;
-    @FXML
-    private TextField userInput;
-    @FXML
-    private Button sendButton;
-
-    private Duke duke;
-
-    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/cat.png"));
-    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/bongoCat.gif"));
-
-    @FXML
-    public void initialize() {
-        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
-    }
-
-    public void setDuke(Duke d) {
-        duke = d;
-    }
-
-    /**
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
-     */
-    @FXML
-    private void handleUserInput() {
-        String input = userInput.getText();
-        String response = duke.getResponse(input);
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(input, userImage),
-                DialogBox.getDukeDialog(response, dukeImage)
-        );
-        if (duke.isExit()) {
-            Platform.exit();
-        }
-        userInput.clear();
-    }
-
-    /**
-     * Shows Duke's welcome message.
-     */
-    public void showWelcome() {
-        String welcome = duke.getWelcome();
-        dialogContainer.getChildren().add(
-                DialogBox.getDukeDialog(welcome, dukeImage)
-        );
-    }
-
-
-}
