+++ src/main/java/manmeowth/command/ListCommand.java
@@ -1,0 +1,21 @@
+package manmeowth.command;
+
+import manmeowth.list.TaskList;
+import manmeowth.storage.ListLoader;
+import manmeowth.ui.Ui;
+
+/**
+ * Represents command to ManMeowth to print list.
+ *
+ * @author WR3nd3
+ */
+public class ListCommand extends Command {
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String execute(TaskList tasks, Ui ui, ListLoader storage) {
+        return ui.showList(tasks.giveList());
+    }
+}
