+++ src/main/java/manmeowth/task/TaskId.java
@@ -1,0 +1,10 @@
+package manmeowth.task;
+
+/**
+ * Identifiers to be used for tasks.
+ *
+ * @author WR3nd3
+ */
+public enum TaskId {
+    T, E, D
+}
