+++ src/main/java/manmeowth/ui/MainWindow.java
@@ -1,0 +1,73 @@
+package manmeowth.ui;
+
+import javafx.animation.PauseTransition;
+import javafx.application.Platform;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+import javafx.util.Duration;
+import manmeowth.ManMeowth;
+
+/**
+ * Controller for MainWindow. Provides the layout for the other controls.
+ */
+public class MainWindow extends AnchorPane {
+    @FXML
+    private ScrollPane scrollPane;
+    @FXML
+    private VBox dialogContainer;
+    @FXML
+    private TextField userInput;
+    @FXML
+    private Button sendButton;
+
+    private ManMeowth manMeowth;
+
+    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/cat.png"));
+    private Image manMeowthImage = new Image(this.getClass().getResourceAsStream("/images/bongo.png"));
+
+    @FXML
+    public void initialize() {
+        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+    }
+
+    public void setManMeowth(ManMeowth d) {
+        manMeowth = d;
+    }
+
+    /**
+     * Creates two dialog boxes, one echoing user input and the other containing ManMeowth's reply
+     * and then appends them to the dialog container. Clears the user input after processing.
+     */
+    @FXML
+    private void handleUserInput() {
+        String input = userInput.getText();
+        String response = manMeowth.getResponse(input);
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(input, userImage),
+                DialogBox.getManMeowthDialog(response, manMeowthImage)
+        );
+        if (manMeowth.isExit()) {
+            PauseTransition delay = new PauseTransition(Duration.seconds(2.5));
+            delay.setOnFinished(exitEvent -> Platform.exit());
+            delay.play();
+        }
+        userInput.clear();
+    }
+
+    /**
+     * Shows ManMeowth's welcome message.
+     */
+    public void showWelcome() {
+        String welcome = manMeowth.getWelcome();
+        dialogContainer.getChildren().add(
+                DialogBox.getManMeowthDialog(welcome, manMeowthImage)
+        );
+    }
+
+
+}
